+++ src/css/style.css
@@ -170,7 +170,8 @@ button {
   width: 100%;
 }
 
-#productSuggestedPrice, #productFinalPrice{
+#productSuggestedPrice,
+#productFinalPrice {
   display: inline;
   margin-right: 10px;
 }
@@ -179,7 +180,7 @@ button {
 }
 #productFinalPrice {
   font-size: 1.5em;
-  color:#525b0f;
+  color: #525b0f;
 }
 
 /* Start cart list card styles */
