+++ src/js/main.js
@@ -1,4 +1,7 @@
 import productList from "./productList.mjs";
-document.addEventListener("DOMContentLoaded", function () {
+import { loadHeaderFooter } from "./utils.mjs";
+
+document.addEventListener("DOMContentLoaded", async function () {
+  await loadHeaderFooter();
   productList(".product-list", "tents");
 });
