+++ src/js/product.js
@@ -2,8 +2,11 @@ import { getParam } from "./utils.mjs";
 import { findProductById } from "./productData.mjs";
 import productDetails from "./productDetails.mjs";
 import { addProductToCart } from "./productDetails.mjs";
+import { loadHeaderFooter } from "./utils.mjs";
+
 const productId = getParam("product");
 productDetails(productId);
+loadHeaderFooter();
 
 // add to cart button event handler
 async function addToCartHandler(e) {
