+++ src/partials/footer.html
@@ -1,0 +1,1 @@
+&copy;NOT a real business
